refactor: Use MockMVC Tests instead of full integration tests
Prior to this change, we used full Integration tests that spin up a full application. That takes a lot of time and needs to allocate a random port.

This change switches to MockMvc test - that's much faster.

Also, this extracts all the test configuration to `BaseIntegrationTest` so more tests can be written.
