+++ pulpogato-rest-fpt/src/test/java/io/github/pulpogato/rest/api/UsersApiIntegrationTest.java
@@ -1,55 +1,12 @@
 package io.github.pulpogato.rest.api;
 
-import io.github.pulpogato.test.ProxyApplication;
-import org.junit.jupiter.api.BeforeEach;
+import io.github.pulpogato.test.BaseIntegrationTest;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInfo;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.client.reactive.ReactorClientHttpConnector;
-import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.reactive.function.client.support.WebClientAdapter;
-import org.springframework.web.service.invoker.HttpServiceProxyFactory;
-import reactor.netty.http.client.HttpClient;
-
-import java.lang.reflect.Method;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest(
-        webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
-        classes = {ProxyApplication.class},
-        properties = {
-                "spring.application.name=pulpogato-demo",
-                "logging.level.org.springframework.web.client.RestTemplate=DEBUG",
-                "logging.level.org.apache.http.wire=DEBUG",
-        }
-)
-class UsersApiIntegrationTest {
-
-    @LocalServerPort
-    int randomServerPort;
-
-    HttpServiceProxyFactory factory;
-
-    @BeforeEach
-    void setUp(TestInfo testInfo) {
-        HttpClient httpClient = HttpClient.create();
-
-        String classPart = testInfo.getTestClass().map(Class::getName).map(name -> name.replace(".", "/")).orElseThrow();
-        String methodPart = testInfo.getTestMethod().map(Method::getName).orElseThrow();
-
-        final var webClient = WebClient.builder()
-                .baseUrl("http://localhost:" + randomServerPort)
-                .clientConnector(new ReactorClientHttpConnector(httpClient))
-                .defaultHeader("TapeName", classPart + "/" + methodPart)
-                .build();
-
-        factory = HttpServiceProxyFactory.builder()
-                .exchangeAdapter(WebClientAdapter.create(webClient))
-                .build();
-    }
+class UsersApiIntegrationTest extends BaseIntegrationTest {
 
     @Test
     void testGetAuthenticatedPublic() {
